requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
aiofiles==24.1.0
    # via aiogram
aiogram==3.17.0
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via aiogram
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==24.3.0
    # via
    #   aiohttp
    #   service-identity
    #   twisted
automat==24.8.1
    # via twisted
build==1.2.2.post1
certifi==2024.12.14
    # via
    #   aiogram
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
colorama==0.4.6
    # via
    #   build
    #   click
constantly==23.10.4
    # via twisted
cryptography==44.0.0
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.2.0
    # via
    #   parsel
    #   scrapy
defusedxml==0.7.1
    # via scrapy
filelock==3.16.1
    # via tldextract
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
hyperlink==21.0.0
    # via twisted
idna==3.10
    # via
    #   hyperlink
    #   requests
    #   tldextract
    #   yarl
incremental==24.7.2
    # via twisted
itemadapter==0.10.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.3.2
    # via scrapy
jmespath==1.0.1
    # via
    #   itemloaders
    #   parsel
lxml==5.3.0
    # via
    #   parsel
    #   scrapy
magic-filter==1.0.12
    # via aiogram
msgspec==0.19.0
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
orjson==3.10.14
packaging==24.2
    # via
    #   build
    #   parsel
    #   scrapy
parsel==1.10.0
    # via
    #   itemloaders
    #   scrapy
pip==24.3.1
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protego==0.4.0
    # via scrapy
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.1
    # via service-identity
pycparser==2.22
    # via cffi
pydantic==2.10.5
    # via aiogram
pydantic-core==2.27.2
    # via pydantic
pydispatcher==2.0.7
    # via scrapy
pyopenssl==25.0.0
    # via scrapy
pyproject-hooks==1.2.0
    # via build
python-dotenv==1.0.1
queuelib==1.7.0
    # via scrapy
requests==2.32.3
    # via
    #   requests-file
    #   tldextract
requests-file==2.1.0
    # via tldextract
scrapy==2.12.0
service-identity==24.2.0
    # via scrapy
setuptools==75.8.0
    # via
    #   incremental
    #   zope-interface
six==1.17.0
    # via transliterate
soupsieve==2.6
tldextract==5.1.3
    # via scrapy
transliterate==1.10.2
twisted==24.11.0
    # via scrapy
typing-extensions==4.12.2
    # via
    #   aiogram
    #   pydantic
    #   pydantic-core
    #   twisted
urllib3==2.3.0
    # via requests
uv==0.5.21
w3lib==2.2.1
    # via
    #   parsel
    #   scrapy
wheel==0.45.1
yarl==1.18.3
    # via aiohttp
zope-interface==7.2
    # via
    #   scrapy
    #   twisted
